Develop backend

See merge request s09-webmobile1-sub2/S09P12A206!316

+++ backend/docker-build-backend.sh
@@ -1,4 +1,4 @@
-docker rm -f sswm-springboot-container
-docker rmi -f ubuntu_application:latest
+if docker ps -a | grep sswm-springboot-container; then docker rm -f sswm-springboot-container; fi
+if docker ps -a | grep app-server_application:latest; then docker rmi -f app-server_application:latest; fi
 cd app-server
 sudo docker-compose up -d
